ライブラリ更新

+++ Common/load/load_mesh.h
@@ -14,10 +14,10 @@ namespace load
     {
     public:
         //Char name_[MaxNameSize];
-        s32 geometry_; /// ジオメトリインデックス
-        s32 material_; /// マテリアルインデックス
-        u16 indexOffset_;
-        u16 numIndices_; /// インデックス数
+        s16 geometry_; /// ジオメトリインデックス
+        s16 material_; /// マテリアルインデックス
+        u32 indexOffset_;
+        u32 numIndices_; /// インデックス数
         lmath::Sphere sphere_; /// 境界球
     };
 }
